🚧 updated front and back
